double -> unsigned long long (working)
store money in internal memory as fraction of whole value (in this case cents).
the size of the fraction is determined by DECIMAL_PLACES.  not a ton of testing
to try out different values for DECIMAL_PLACES, but it is meant to be flexible in
the future.  Previous approach used doubles to store money, but this was too buggy